python-api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
pandas==2.0.3
faiss-cpu==1.7.4
torch==2.0.1
transformers==4.33.2
groq==0.4.1
networkx==3.1
scikit-learn==1.3.0
sentence-transformers==2.2.2
spacy==3.6.1
scispacy==0.5.3
joblib==1.3.2
gunicorn==21.2.0
